requirements.txt
graphics.py==5.0.1.post1
numpy==1.22.2
